+++ src/components/common/OutlinedButton.tsx
@@ -19,7 +19,7 @@ export const OutlinedButton = ({
   ref,
   type = "button",
 }: OutlinedButtonProps) => {
-  const buttonClassName = ` flex rounded-md justify-center items-center gap-2 w-full h-full p-2 !outline-0 border border-mainBorder dark:border-mainBorderDark bg-outlinedButtonBg dark:bg-outlinedButtonBgDark hover:bg-outlinedButtonBgHover dark:hover:bg-outlinedButtonBgHoverDark text-primaryText dark:text-primaryTextDark dark:stroke-grayIconDark dark:fill-grayIconDark fill-grayIcon stroke-grayIcon ${className}`;
+  const buttonClassName = `text-md flex rounded-md justify-center items-center gap-2 w-full h-full p-2 !outline-0 border border-mainBorder dark:border-mainBorderDark bg-outlinedButtonBg dark:bg-outlinedButtonBgDark hover:bg-outlinedButtonBgHover dark:hover:bg-outlinedButtonBgHoverDark text-primaryText dark:text-primaryTextDark dark:stroke-grayIconDark dark:fill-grayIconDark fill-grayIcon stroke-grayIcon ${className}`;
 
   return (
     <button onClick={handleClick} className={buttonClassName} ref={ref}>