+++ src/components/views/orders/OrdersDateRange.tsx
@@ -47,7 +47,7 @@ export const OrdersDateRange = ({
           ref={startDatePickerRef}
           selected={startDateObj}
           onChange={handleStartDateChange}
-          className="pl-3 max-w-[10rem] min-w-[10rem] p-2 text-sm 1xl:text-base bg-inputBg dark:bg-inputBgDark  hover:dark:bg-inputBgHoverDark hover:bg-InputBgHover w-full h-full border rounded-md border-inputBorder dark:border-inputBorderDark  text-primaryText placeholder-secondaryText dark:placeholder-secondaryTextDark dark:text-primaryTextDark hover:dark:border-inputBorderHoverDark hover:border-inputBorderHover transition"
+          className="pl-3 md:max-w-[10rem] md:min-w-[10rem] p-2 text-sm 1xl:text-base bg-inputBg dark:bg-inputBgDark  hover:dark:bg-inputBgHoverDark hover:bg-InputBgHover w-full h-full border rounded-md border-inputBorder dark:border-inputBorderDark  text-primaryText placeholder-secondaryText dark:placeholder-secondaryTextDark dark:text-primaryTextDark hover:dark:border-inputBorderHoverDark hover:border-inputBorderHover transition"
         />
         <div
           // @ts-ignore
@@ -62,7 +62,7 @@ export const OrdersDateRange = ({
           ref={endDatePickerRef}
           selected={endDateObj}
           onChange={handleEndDateChange}
-          className="pl-3 max-w-[10rem] min-w-[10rem] p-2 text-sm 1xl:text-base bg-inputBg dark:bg-inputBgDark  hover:dark:bg-inputBgHoverDark hover:bg-InputBgHover w-full h-full border rounded-md border-inputBorder dark:border-inputBorderDark  text-primaryText placeholder-secondaryText dark:placeholder-secondaryTextDark dark:text-primaryTextDark hover:dark:border-inputBorderHoverDark hover:border-inputBorderHover transition"
+          className="pl-3 md:max-w-[10rem] md:min-w-[10rem] p-2 text-sm 1xl:text-base bg-inputBg dark:bg-inputBgDark  hover:dark:bg-inputBgHoverDark hover:bg-InputBgHover w-full h-full border rounded-md border-inputBorder dark:border-inputBorderDark  text-primaryText placeholder-secondaryText dark:placeholder-secondaryTextDark dark:text-primaryTextDark hover:dark:border-inputBorderHoverDark hover:border-inputBorderHover transition"
         />
         <div
           // @ts-ignore