+++ src/layout/sideMenu/SideMenuMobile.tsx
@@ -15,6 +15,8 @@ import { ProductsIcon } from "../../assets/icons/ProductsIcon";
 import { Select } from "../../components/forms/Select";
 import { useTheme } from "next-themes";
 import { useSession } from "../../hooks/auth/useSession";
+import { Link } from "../../i18n/navigation";
+import { useLocale } from "next-intl";
 
 interface SideMenuMobileProps {
   isMobileMenuOpen: boolean;
@@ -29,15 +31,18 @@ export const SideMenuMobile = ({
   const { session } = useSession();
   const { theme } = useTheme();
 
+  const locale = useLocale();
+console.log(locale)
   return (
     <div
-      className={` overflow-auto flex fixed xl:hidden flex-col bg-primaryBg  border-r-[1px] border-mainBorder dark:border-mainBorderDark dark:bg-primaryBgDark white top-[5rem] xl:top-[4rem] 2xl:top-[5rem] mb-[3rem] left-0 items-center transform transition-transform ease-in-out  ${
+      className={`z-50 overflow-auto flex fixed xl:hidden flex-col justify-between bg-primaryBg  border-r-[1px] border-mainBorder dark:border-mainBorderDark dark:bg-primaryBgDark white top-[4.5rem] xl:top-[4rem] 2xl:top-[4.5rem] mb-[2.5rem] left-0 items-center transform transition-transform ease-in-out  ${
         isMobileMenuOpen ? "translate-x-0" : "-translate-x-full"
       }
       ${theme === "prismatic" && "backdrop-blur-xl"}
       `}
-      style={{ height: "calc(100% - 5rem)" }}
+      style={{ height: "calc(100% - 4.5rem)" }}
     >
+     
       <div className="px-4 xl:px-6 pt-2 pr-6  transition w-[16rem] pb-2">
         <MenuCategory title="Main" />
         <MenuItem title="Dashboard" icon={<DashboardIcon />} path="/" />
@@ -61,39 +66,53 @@ export const SideMenuMobile = ({
         <MenuItem title="Scatter" icon={<ScatterIcon />} path="/scatter" />
         <MenuItem title="Line" icon={<LineIcon />} path="/line" />
       </div>
-      <div className="w-full border-t-0 dark:border-mainBorderDark px-4 pt-8 mb-6 ">
-        {!session?.isLoggedIn && (
-          <button
-            onClick={() => {
-              onLoginButtonClick();
-              toggleMobileMenu();
-            }}
-            className="block hover:bg-navbarButtonBgHover hover:dark:bg-navbarButtonBgHoverDark xl:hidden mt-auto mb-8 rounded-xl w-full h-10 flex justify-center items-center font-medium border border-mainColor dark:border-mainColorDark text-primaryText dark:text-primaryTextDark bg-[rgb(255,255,255,0.02)] dark:hover:bg-[rgb(255,255,255,0.06)] mt-12"
-          >
-            Sign In
-          </button>
-        )}
-        <div className="flex xl:hidden justify-center gap-2 items-center mx-2">
-          <label className="mr-2 text-primaryText dark:text-primaryTextDark">
-            Language:
-          </label>
-          <Select>
-            <option value="english">English</option>
-            <option value="polish">Polish</option>
-          </Select>
+      <div className="w-full">
+        <div className="w-full border-t-0 dark:border-mainBorderDark px-4 pt-8 mb-6 ">
+          {!session?.isLoggedIn && (
+            <button
+              onClick={() => {
+                onLoginButtonClick();
+                toggleMobileMenu();
+              }}
+              className="block hover:bg-navbarButtonBgHover hover:dark:bg-navbarButtonBgHoverDark xl:hidden mt-auto mb-8 rounded-xl w-full h-10 flex justify-center items-center font-medium border border-mainColor dark:border-mainColorDark text-primaryText dark:text-primaryTextDark bg-[rgb(255,255,255,0.02)] dark:hover:bg-[rgb(255,255,255,0.06)] mt-12"
+            >
+              Sign In
+            </button>
+          )}
+          <div className="flex xl:hidden justify-center gap-2 items-center mx-2">
+            <label className="mr-2 text-primaryText dark:text-primaryTextDark">
+              Language:
+            </label>
+            <div className="flex border border-mainBorder dark:border-mainBorderDark rounded-md">
+              <Link
+                href="/"
+                locale="en"
+                className={`${locale === "en" ? "border border-mainBorder dark:border-mainBorderDark" : "border border-[rgb(0,0,0,0)] dark:border-[rgb(0,0,0,0)]" }  text-primaryText dark:text-primaryTextDark h-10 cursor-pointer px-4 hover:bg-dropdownBgHover hover:dark:bg-dropdownBgHoverDark py-2 flex justify-between`}
+              >
+                EN
+              </Link>
+              <Link
+                href="/"
+                locale="pl"
+                className={`${locale === "pl" ? "border border-mainBorder dark:border-mainBorderDark" : "border border-[rgb(0,0,0,0)] dark:border-[rgb(0,0,0,0)]" } text-primaryText dark:text-primaryTextDark h-10 cursor-pointer px-4 hover:bg-dropdownBgHover hover:dark:bg-dropdownBgHoverDark py-2 flex justify-between`}
+              >
+                PL
+              </Link>{" "}
+            </div>
+          </div>
         </div>
+        <a
+          href="https://github.com/matt765/spireflow"
+          target="_blank"
+          rel="noreferrer"
+          className="cursor-pointer hover:bg-navbarButtonBgHover hover:dark:bg-navbarButtonBgHoverDark min-h-[3.6rem] max-h-[3.6rem] w-full border-t-2 border-mainBorder dark:border-mainBorderDark  text-center flex justify-center items-center gap-2 stroke-grayIcon fill-grayIcon dark:stroke-grayIconDark dark:fill-grayIconDark"
+        >
+          <GithubIcon />
+          <div className="text-primaryText dark:text-primaryTextDark">
+            GitHub Repository
+          </div>
+        </a>
       </div>
-      <a
-        href="https://github.com/matt765/spireflow"
-        target="_blank"
-        rel="noreferrer"
-        className="cursor-pointer hover:bg-navbarButtonBgHover hover:dark:bg-navbarButtonBgHoverDark min-h-[3.6rem] max-h-[3.6rem] w-full border-t-2 border-mainBorder dark:border-mainBorderDark  text-center flex justify-center items-center gap-2 stroke-grayIcon fill-grayIcon dark:stroke-grayIconDark dark:fill-grayIconDark"
-      >
-        <GithubIcon />
-        <div className="text-primaryText dark:text-primaryTextDark">
-          GitHub Repository
-        </div>
-      </a>
     </div>
   );
 };